Extract FileCopier from tooling to reuse in msbuild so we can clone when we copy
