+++ tools/common/Target.cs
@@ -98,14 +98,6 @@ namespace Xamarin.Bundler {
 #endif
 		}
 
-		[DllImport (Constants.libSystemLibrary, SetLastError = true, EntryPoint = "strerror")]
-		static extern IntPtr _strerror (int errno);
-
-		internal static string strerror (int errno)
-		{
-			return Marshal.PtrToStringAuto (_strerror (errno));
-		}
-
 		[DllImport (Constants.libSystemLibrary, SetLastError = true)]
 		static extern string realpath (string path, IntPtr zero);
 
@@ -116,7 +108,7 @@ namespace Xamarin.Bundler {
 				return rv;
 
 			var errno = Marshal.GetLastWin32Error ();
-			ErrorHelper.Warning (54, "Unable to canonicalize the path '{0}': {1} ({2}).", path, strerror (errno), errno);
+			ErrorHelper.Warning (54, "Unable to canonicalize the path '{0}': {1} ({2}).", path, FileCopier.strerror (errno), errno);
 			return path;
 		}
